updates to new hash based parser and correction to Locus_GI datatype in database creation code

+++ database/Chromosome17.sql
@@ -12,7 +12,7 @@ CREATE TABLE Chromosome_Locations
 
 CREATE TABLE Loci
 (	Genbank_Accession		VARCHAR(15)	DEFAULT NULL,
-	Locus_GI				SMALLINT	NOT NULL,
+	Locus_GI				INT			NOT NULL,
 	DNA_seq					LONGTEXT	NOT NULL,
 	Product_Name			TEXT		NOT NULL,
 	CDS_translated 			TEXT		NOT NULL,
